refactor: split out modal and image tests and add caption/credits to snaps
