+++ packages/image/__tests__/android/modal-image-with-style.android.test.js
@@ -1,0 +1,3 @@
+import shared from "../modal-shared-with-style.native";
+
+shared();
