+++ packages/image/__tests__/android/modal-image.android.test.js
@@ -1,0 +1,3 @@
+import shared from "../modal-shared.native";
+
+shared();
