+++ packages/image/__tests__/ios/modal-image-with-style.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../modal-shared-with-style.native";
+
+shared();
