+++ packages/image/__tests__/ios/modal-image.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../modal-shared.native";
+
+shared();
