+++ packages/image/__tests__/modal-shared-with-style.native.js
@@ -10,7 +10,14 @@ import {
 } from "@times-components/jest-serializer";
 import { iterator } from "@times-components/test-utils";
 import "./mocks";
-import Image from "../src";
+import { ModalImage } from "../src";
+
+const props = {
+  aspectRatio: 3 / 2,
+  caption: "Test caption",
+  credits: "Test credits",
+  uri: "http://example.com/image.jpg"
+};
 
 export default () => {
   addSerializers(
@@ -25,11 +32,9 @@ export default () => {
 
   const tests = [
     {
-      name: "default image",
+      name: "default modal",
       test: () => {
-        const testInstance = TestRenderer.create(
-          <Image aspectRatio={3 / 2} uri="http://example.com/image.jpg" />
-        );
+        const testInstance = TestRenderer.create(<ModalImage {...props} />);
 
         expect(testInstance).toMatchSnapshot();
       }
